source_db_api/requirements.txt
fastapi==0.110.*
pydantic==2.6.*
uvicorn==0.29.*
sqlalchemy==2.0.*
python-dotenv==1.0.*
psycopg2-binary==2.9.*